@GetMapping("/change-locale") moved from IndexController.java to LocaleController.java.
