+++ src/main/resources/languages/messages_pl.properties
@@ -1,6 +1,6 @@
 # Navbar links
 navbar.home=StronaPL
-navbar.lang.chooseLanguage=J?zyk
+navbar.lang.chooseLanguage=Język
 navbar.lang.pl=polski
 navbar.lang.en=angielski
